@@ -1,8 +1,11 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\Models\CastMember;
 use Illuminate\Database\Seeder;
 
-class CategorySeeder extends Seeder
+class CastMemberSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,6 +14,6 @@ class CategorySeeder extends Seeder
      */
     public function run()
     {
-        factory(\App\Models\Category::class, 100)->create();
+        CastMember::factory()->count(100)->create();
     }
 }
