@@ -1,8 +1,11 @@
 <?php
 
+namespace Database\Seeders;
+
+use App\Models\Category;
 use Illuminate\Database\Seeder;
 
-class CastMemberSeeder extends Seeder
+class CategorySeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -11,6 +14,6 @@ class CastMemberSeeder extends Seeder
      */
     public function run()
     {
-        factory(\App\Models\CastMember::class, 100)->create();
+        Category::factory()->count(100)->create();
     }
 }
