@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Models\Category;
 use App\Models\Genre;
 use Illuminate\Database\Seeder;
@@ -12,7 +14,7 @@ class GenreSeeder extends Seeder
     public function run(): void
     {
         $categories = Category::all();
-        factory(Genre::class, 100)
+        Genre::factory()->count(100)
             ->create()
             ->each(function (Genre $genre) use ($categories) {
                 $genre->categories()->sync($categories->random(5));
