@@ -8,7 +8,7 @@ use App\Models\Category;
 use App\Models\Genre;
 use App\Models\Video;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
-use Illuminate\Foundation\Testing\TestResponse;
+use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
 use Tests\Traits\TestResources;
 use Tests\Traits\TestSaves;
@@ -26,10 +26,10 @@ abstract class BaseVideoControllerTestCase extends TestCase
     protected function setUp(): void
     {
         parent::setUp();
-        $this->video = factory(Video::class)->create(['opened' => false]);
-        $category = factory(Category::class)->create();
-        $genre = factory(Genre::class)->create();
-        $castMember = factory(CastMember::class)->create();
+        $this->video = Video::factory()->create(['opened' => false]);
+        $category = Category::factory()->create();
+        $genre = Genre::factory()->create();
+        $castMember = CastMember::factory()->create();
         $genre->categories()->sync($category);
         $this->sendData = [
             'title' => 'title',
