@@ -3,41 +3,51 @@
 namespace App\Models;
 
 use App\ModelFilters\CategoryFilter;
-use App\Models\Traits\UuidModel;
+use App\Models\UuidModel;
 use EloquentFilter\Filterable;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 /**
  * App\Models\Category.
  *
- * @property string                          $id
- * @property string                          $name
- * @property null|string                     $description
- * @property bool                            $is_active
- * @property null|\Illuminate\Support\Carbon $deleted_at
- * @property null|\Illuminate\Support\Carbon $created_at
- * @property null|\Illuminate\Support\Carbon $updated_at
- *
+ * @property string                                                       $id
+ * @property string                                                       $name
+ * @property null|string                                                  $description
+ * @property bool                                                         $is_active
+ * @property null|\Illuminate\Support\Carbon                              $deleted_at
+ * @property null|\Illuminate\Support\Carbon                              $created_at
+ * @property null|\Illuminate\Support\Carbon                              $updated_at
+ * @property \App\Models\Genre[]|\Illuminate\Database\Eloquent\Collection $genres
+ * @property null|int                                                     $genres_count
+ * @method static \Illuminate\Database\Eloquent\Builder|Category filter(array $input = [], $filter = null)
  * @method static \Illuminate\Database\Eloquent\Builder|Category newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder|Category newQuery()
  * @method static \Illuminate\Database\Query\Builder|Category onlyTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder|Category paginateFilter($perPage = null, $columns = [], $pageName = 'page', $page = null)
  * @method static \Illuminate\Database\Eloquent\Builder|Category query()
+ * @method static \Illuminate\Database\Eloquent\Builder|Category simplePaginateFilter(?int $perPage = null, ?int $columns = [], ?int $pageName = 'page', ?int $page = null)
+ * @method static \Illuminate\Database\Eloquent\Builder|Category whereBeginsWith(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereCreatedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereDeletedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Category whereEndsWith(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereId($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereIsActive($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Category whereLike(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereName($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Category whereUpdatedAt($value)
  * @method static \Illuminate\Database\Query\Builder|Category withTrashed()
  * @method static \Illuminate\Database\Query\Builder|Category withoutTrashed()
  * @mixin \Eloquent
+ * @method static \Database\Factories\CategoryFactory factory(...$parameters)
  */
 class Category extends UuidModel
 {
     use SoftDeletes;
     use Filterable;
+    use HasFactory;
 
     protected $fillable = ['name', 'description', 'is_active'];
 
