@@ -3,66 +3,74 @@
 namespace App\Models;
 
 use App\Models\Traits\UploadFiles;
-use App\Models\Traits\UuidModel;
+use App\Models\UuidModel;
 use EloquentFilter\Filterable;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 /**
  * App\Models\Video.
  *
- * @property string                                                          $id
- * @property string                                                          $title
- * @property string                                                          $description
- * @property int                                                             $year_launched
- * @property bool                                                            $opened
- * @property string                                                          $rating
- * @property int                                                             $duration
- * @property null|string                                                     $video_file
- * @property null|\Illuminate\Support\Carbon                                 $deleted_at
- * @property null|\Illuminate\Support\Carbon                                 $created_at
- * @property null|\Illuminate\Support\Carbon                                 $updated_at
- * @property \App\Models\Category[]|\Illuminate\Database\Eloquent\Collection $categories
- * @property null|int                                                        $categories_count
- * @property \App\Models\Genre[]|\Illuminate\Database\Eloquent\Collection    $genres
- * @property null|int                                                        $genres_count
- *
+ * @property string                                                            $id
+ * @property string                                                            $title
+ * @property string                                                            $description
+ * @property int                                                               $year_launched
+ * @property bool                                                              $opened
+ * @property string                                                            $rating
+ * @property int                                                               $duration
+ * @property null|string                                                       $video_file
+ * @property null|string                                                       $thumb_file
+ * @property null|string                                                       $trailer_file
+ * @property null|string                                                       $banner_file
+ * @property null|\Illuminate\Support\Carbon                                   $deleted_at
+ * @property null|\Illuminate\Support\Carbon                                   $created_at
+ * @property null|\Illuminate\Support\Carbon                                   $updated_at
+ * @property \App\Models\CastMember[]|\Illuminate\Database\Eloquent\Collection $castMembers
+ * @property null|int                                                          $cast_members_count
+ * @property \App\Models\Category[]|\Illuminate\Database\Eloquent\Collection   $categories
+ * @property null|int                                                          $categories_count
+ * @property \App\Models\Genre[]|\Illuminate\Database\Eloquent\Collection      $genres
+ * @property null|int                                                          $genres_count
+ * @property null|string                                                       $banner_file_url
+ * @property null|string                                                       $thumb_file_url
+ * @property null|string                                                       $trailer_file_url
+ * @property null|string                                                       $video_file_url
+ * @method static \Illuminate\Database\Eloquent\Builder|Video filter(array $input = [], $filter = null)
  * @method static \Illuminate\Database\Eloquent\Builder|Video newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder|Video newQuery()
  * @method static \Illuminate\Database\Query\Builder|Video onlyTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder|Video paginateFilter($perPage = null, $columns = [], $pageName = 'page', $page = null)
  * @method static \Illuminate\Database\Eloquent\Builder|Video query()
+ * @method static \Illuminate\Database\Eloquent\Builder|Video simplePaginateFilter(?int $perPage = null, ?int $columns = [], ?int $pageName = 'page', ?int $page = null)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereBannerFile($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereBeginsWith(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereCreatedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereDeletedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereDescription($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereDuration($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereEndsWith(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereLike(string $column, string $value, string $boolean = 'and')
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereOpened($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereRating($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereThumbFile($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereTitle($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Video whereTrailerFile($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereUpdatedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereVideoFile($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Video whereYearLaunched($value)
  * @method static \Illuminate\Database\Query\Builder|Video withTrashed()
  * @method static \Illuminate\Database\Query\Builder|Video withoutTrashed()
  * @mixin \Eloquent
- *
- * @property null|string $thumb_file
- * @property null|string $trailer_file
- * @property null|string $banner_file
- * @property null|string $banner_file_url
- * @property null|string $thumb_file_url
- * @property null|string $trailer_file_url
- * @property null|string $video_file_url
- *
- * @method static \Illuminate\Database\Eloquent\Builder|Video whereBannerFile($value)
- * @method static \Illuminate\Database\Eloquent\Builder|Video whereThumbFile($value)
- * @method static \Illuminate\Database\Eloquent\Builder|Video whereTrailerFile($value)
+ * @method static \Database\Factories\VideoFactory factory(...$parameters)
  */
 class Video extends UuidModel
 {
     use SoftDeletes;
     use UploadFiles;
     use Filterable;
+    use HasFactory;
 
     public const RATING_LIST = ['L', '10', '12', '14', '16', '18'];
 
