@@ -1,13 +1,29 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Models\CastMember;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(CastMember::class, function (Faker $faker) {
-    return [
-        'name' => $faker->firstName,
-        'type' => $faker->randomElement([CastMember::TYPE_DIRECTOR, CastMember::TYPE_ACTOR]),
-    ];
-});
+class CastMemberFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = CastMember::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->firstName,
+            'type' => $this->faker->randomElement([CastMember::TYPE_DIRECTOR, CastMember::TYPE_ACTOR]),
+        ];
+    }
+}
