@@ -1,17 +1,33 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Models\Video;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Video::class, function (Faker $faker) {
-    return [
-        'title' => $faker->sentence(3),
-        'description' => $faker->sentence(10),
-        'year_launched' => rand(1895, 2020),
-        'opened' => rand(0, 1),
-        'rating' => $faker->randomElement(Video::RATING_LIST),
-        'duration' => rand(1, 30),
-    ];
-});
+class VideoFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Video::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'title' => $this->faker->sentence(3),
+            'description' => $this->faker->sentence(10),
+            'year_launched' => rand(1895, 2020),
+            'opened' => rand(0, 1),
+            'rating' => $this->faker->randomElement(Video::RATING_LIST),
+            'duration' => rand(1, 30),
+        ];
+    }
+}
